closetx_app/requirements.txt
flask
pyjwt
boto3
mysql-connector-python
pillow
opencv-python
